chore: merge functionality from util.js
